rename plugin.config.ts => plugin.ts
