+++ src/app/frontend/common/services/global/plugin.ts
